pyproject.toml
[tool.poetry]
name = "end4train"
version = "0.1.0"
description = ""
authors = ["Vilém Pecen <pecenvilem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
pandas = "^2.2.2"
pyside6 = "^6.8.0.2"
pyqtgraph = "^0.13.7"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pyarrow = "^17.0.0"


[tool.poetry.group.build.dependencies]
pyinstaller = "^6.11.0"
kaitaistruct = {git = "https://github.com/kaitai-io/kaitai_struct_python_runtime.git"}
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
